Create Thanksgiving
                                                                                                    Thanksgiving plan.

This year is a little different form all the years before because of the corona virus. I am not sure if we are ever going to have gatherings and holidays with family and friends like we used to do before. this pandemic has changed people forever. Sad but true, things will not go back to normal even if this pandemic ends.
Here what I am going to do this thanksgiving for the sake of the kids to feel that we still can celebrate even with this current situation. 
1-	Buying a big chicken and prepare it just like preparing a turkey accept that the kids do not like turkey so I have to make the same thing but with chicken. 
2-	Making sweets. like baklava or kanafeh. No one likes pies in my house. 
3-	Mashed potatoes is a must
4-	Stuffing
5-	Pomegranate juice. 
It is going to be very small but it is going to be a nice celebration